requirements.txt
PyMuPDF>=1.23.0
pdfminer.six>=20221105
python-pptx>=0.6.21
pikepdf>=8.0.0
lxml>=4.9.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=9.5.0
